Include GlmSharp and refactor float to double throughout Engine
These are only refactorings of existing content, and should
exhibit the same behavior as before the commit (except with
higher precision now). Using GlmSharp, we can use proper
matrix-vector mathematics and also use different precision
for vector types, such as double.

+++ engine/anatomy/joints/HingeJoint.cs
@@ -1,5 +1,19 @@
-using System.Collections;
-using System.Numerics;
+/*
+ * Copyright (C) 2021 Freedom of Form Foundation, Inc.
+ * 
+ * This program is free software; you can redistribute it and/or
+ * modify it under the terms of the GNU General Public License, version 2 (GPLv2) as published by the Free Software Foundation.
+ * 
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License, version 2 (GPLv2) for more details.
+ * 
+ * You should have received a copy of the GNU General Public License, version 2 (GPLv2)
+ * along with this program; if not, write to the Free Software
+ * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
+ */
+
 using FreedomOfFormFoundation.AnatomyEngine.Geometry;
 using FreedomOfFormFoundation.AnatomyEngine.Calculus;
 
@@ -14,12 +28,12 @@ namespace FreedomOfFormFoundation.AnatomyEngine.Anatomy.Joints
 			articularSurface = new SymmetricCylinder(centerLine, radius);
 		}
 		
-		public HingeJoint(LineSegment centerLine, RaytraceableFunction1D radius, ContinuousMap<float, float> startAngle, ContinuousMap<float, float> endAngle)
+		public HingeJoint(LineSegment centerLine, RaytraceableFunction1D radius, ContinuousMap<double, double> startAngle, ContinuousMap<double, double> endAngle)
 		{
 			articularSurface = new SymmetricCylinder(centerLine, radius, startAngle, endAngle);
 		}
 		
-		//public HingeJoint(LineSegment centerLine, float radius)
+		//public HingeJoint(LineSegment centerLine, double radius)
 		//{
 		//	articularSurface = new SymmetricCylinder(centerLine, radius);
 		//}
